% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randSeed.R
\name{randSeed}
\alias{randSeed}
\title{Generate random seed}
\usage{
randSeed()
}
\description{
This function randomly generate a see from the lower digit on the local machine's clock. The function requires 
input parameters.
}
\examples{
randSeed()
}
